+++ src/features/projects/components/ProjectDetails/ProjectContributions.tsx
@@ -1,9 +1,7 @@
 import { Markdown } from "~/components/ui/Markdown";
 import { Heading } from "~/components/ui/Heading";
-import { LinkBox } from "./LinkBox";
-import { FileCode, Github, Globe, type LucideIcon } from "lucide-react";
-import { createElement } from "react";
 import { type Application } from "~/features/applications/types";
+import Links from "./Links";
 
 type Props = { isLoading: boolean; project?: Application };
 
@@ -19,29 +17,7 @@ export default function ProjectContributions({ isLoading, project }: Props) {
             {project?.contributionDescription}
           </Markdown>
         </div>
-        <div className="md:w-1/3">
-          <LinkBox
-            label="Contribution Links"
-            links={project?.contributionLinks}
-            renderItem={(link) => {
-              const icon: LucideIcon | undefined = {
-                CONTRACT_ADDRESS: FileCode,
-                GITHUB_REPO: Github,
-                OTHER: Globe,
-              }[link.type];
-              return (
-                <>
-                  {createElement(icon ?? "div", {
-                    className: "w-4 h-4 mt-1",
-                  })}
-                  <div className="flex-1 truncate" title={link.description}>
-                    {link.description}
-                  </div>
-                </>
-              );
-            }}
-          />
-        </div>
+        <Links label="Contribution Links" links={project?.contributionLinks} />
       </div>
     </>
   );
